remove `scene` indirection in re_space_view_spatial
