+++ crates/re_space_view_spatial/src/contexts/annotation_context.rs
