+++ crates/re_space_view_spatial/src/contexts/depth_offsets.rs
