+++ crates/re_space_view_spatial/src/contexts/non_interactive_entities.rs
