+++ crates/re_space_view_spatial/src/lib.rs
@@ -3,11 +3,13 @@
 //! Space Views that show entities in a 2D or 3D spatial relationship.
 
 mod axis_lines;
+mod contexts;
 mod eye;
 mod instance_hash_conversions;
 mod mesh_cache;
 mod mesh_loader;
-mod scene;
+mod parts;
+mod picking;
 mod space_camera_3d;
 mod space_view_class;
 mod ui;
