+++ crates/re_space_view_spatial/src/parts/lines3d.rs
@@ -1,4 +1,4 @@
-use re_components::{ColorRGBA, Component as _, InstanceKey, LineStrip2D, Radius};
+use re_components::{ColorRGBA, Component as _, InstanceKey, LineStrip3D, Radius};
 use re_data_store::EntityPath;
 use re_query::{EntityView, QueryError};
 use re_renderer::Size;
@@ -8,23 +8,21 @@ use re_viewer_context::{
 };
 
 use crate::{
-    scene::{
-        contexts::{SpatialSceneContext, SpatialSceneEntityContext},
-        parts::entity_iterator::process_entity_views,
-    },
+    contexts::{SpatialSceneContext, SpatialSceneEntityContext},
+    parts::entity_iterator::process_entity_views,
     SpatialSpaceView,
 };
 
 use super::{picking_id_from_instance_key, SpatialScenePartData, SpatialSpaceViewState};
 
 #[derive(Default)]
-pub struct Lines2DPart(SpatialScenePartData);
+pub struct Lines3DPart(SpatialScenePartData);
 
-impl Lines2DPart {
+impl Lines3DPart {
     fn process_entity_view(
         &mut self,
         _query: &SceneQuery<'_>,
-        ent_view: &EntityView<LineStrip2D>,
+        ent_view: &EntityView<LineStrip3D>,
         ent_path: &EntityPath,
         ent_context: &SpatialSceneEntityContext<'_>,
     ) -> Result<(), QueryError> {
@@ -32,14 +30,14 @@ impl Lines2DPart {
 
         let mut line_builder = ent_context.shared_render_builders.lines();
         let mut line_batch = line_builder
-            .batch("lines 2d")
+            .batch("lines 3D")
             .depth_offset(ent_context.depth_offset)
             .world_from_obj(ent_context.world_from_obj)
             .outline_mask_ids(ent_context.highlight.overall)
             .picking_object_id(re_renderer::PickingLayerObjectId(ent_path.hash64()));
 
         let visitor = |instance_key: InstanceKey,
-                       strip: LineStrip2D,
+                       strip: LineStrip3D,
                        color: Option<ColorRGBA>,
                        radius: Option<Radius>| {
             // TODO(andreas): support class ids for lines
@@ -52,7 +50,7 @@ impl Lines2DPart {
                 annotation_info.color(color.map(move |c| c.to_array()).as_ref(), default_color);
 
             let lines = line_batch
-                .add_strip_2d(strip.0.into_iter().map(|v| v.into()))
+                .add_strip(strip.0.into_iter().map(|v| v.into()))
                 .color(color)
                 .radius(radius)
                 .picking_instance_id(picking_id_from_instance_key(instance_key));
@@ -69,7 +67,7 @@ impl Lines2DPart {
                 strip
                     .map_or(Vec::new(), |strip| strip.0)
                     .into_iter()
-                    .map(|pt| glam::vec3(pt.x(), pt.y(), 0.0))
+                    .map(|pt| pt.into())
             }),
             ent_context.world_from_obj,
         );
@@ -78,10 +76,10 @@ impl Lines2DPart {
     }
 }
 
-impl ViewPartSystem<SpatialSpaceView> for Lines2DPart {
+impl ViewPartSystem<SpatialSpaceView> for Lines3DPart {
     fn archetype(&self) -> ArchetypeDefinition {
         vec1::vec1![
-            LineStrip2D::name(),
+            LineStrip3D::name(),
             InstanceKey::name(),
             ColorRGBA::name(),
             Radius::name(),
@@ -96,9 +94,9 @@ impl ViewPartSystem<SpatialSpaceView> for Lines2DPart {
         scene_context: &SpatialSceneContext,
         highlights: &SpaceViewHighlights,
     ) -> Vec<re_renderer::QueueableDrawData> {
-        re_tracing::profile_scope!("Lines2DPart");
+        re_tracing::profile_scope!("Lines3DPart");
 
-        process_entity_views::<LineStrip2D, 4, _>(
+        process_entity_views::<LineStrip3D, 4, _>(
             ctx,
             query,
             scene_context,
@@ -106,6 +104,9 @@ impl ViewPartSystem<SpatialSpaceView> for Lines2DPart {
             scene_context.depth_offsets.points,
             self.archetype(),
             |_ctx, ent_path, entity_view, ent_context| {
+                scene_context
+                    .num_3d_primitives
+                    .fetch_add(1, std::sync::atomic::Ordering::Relaxed);
                 self.process_entity_view(query, &entity_view, ent_path, ent_context)
             },
         );
