+++ crates/re_space_view_spatial/src/parts/points3d.rs
@@ -1,5 +1,5 @@
 use re_components::{
-    ClassId, ColorRGBA, Component, InstanceKey, KeypointId, Label, Point2D, Radius,
+    ClassId, ColorRGBA, Component as _, InstanceKey, KeypointId, Label, Point3D, Radius,
 };
 use re_data_store::{EntityPath, InstancePathHash};
 use re_query::{EntityView, QueryError};
@@ -9,11 +9,9 @@ use re_viewer_context::{
 };
 
 use crate::{
-    scene::{
-        contexts::{SpatialSceneContext, SpatialSceneEntityContext},
-        load_keypoint_connections,
-        parts::entity_iterator::process_entity_views,
-        UiLabel, UiLabelTarget,
+    contexts::{SpatialSceneContext, SpatialSceneEntityContext},
+    parts::{
+        entity_iterator::process_entity_views, load_keypoint_connections, UiLabel, UiLabelTarget,
     },
     SpatialSpaceView,
 };
@@ -23,13 +21,13 @@ use super::{
     SpatialScenePartData, SpatialSpaceViewState,
 };
 
-pub struct Points2DPart {
+pub struct Points3DPart {
     /// If the number of points in the batch is > max_labels, don't render point labels.
     pub max_labels: usize,
     pub data: SpatialScenePartData,
 }
 
-impl Default for Points2DPart {
+impl Default for Points3DPart {
     fn default() -> Self {
         Self {
             max_labels: 10,
@@ -38,12 +36,13 @@ impl Default for Points2DPart {
     }
 }
 
-impl Points2DPart {
+impl Points3DPart {
     fn process_labels<'a>(
-        entity_view: &'a EntityView<Point2D>,
+        entity_view: &'a EntityView<Point3D>,
         instance_path_hashes: &'a [InstancePathHash],
         colors: &'a [egui::Color32],
         annotation_infos: &'a [ResolvedAnnotationInfo],
+        world_from_obj: glam::Affine3A,
     ) -> Result<impl Iterator<Item = UiLabel> + 'a, QueryError> {
         let labels = itertools::izip!(
             annotation_infos.iter(),
@@ -59,7 +58,9 @@ impl Points2DPart {
                     (Some(point), Some(label)) => Some(UiLabel {
                         text: label,
                         color: *color,
-                        target: UiLabelTarget::Point2D(egui::pos2(point.x, point.y)),
+                        target: UiLabelTarget::Position3D(
+                            world_from_obj.transform_point3(point.into()),
+                        ),
                         labeled_instance: *labeled_instance,
                     }),
                     _ => None,
@@ -72,7 +73,7 @@ impl Points2DPart {
     fn process_entity_view(
         &mut self,
         query: &SceneQuery<'_>,
-        ent_view: &EntityView<Point2D>,
+        ent_view: &EntityView<Point3D>,
         ent_path: &EntityPath,
         ent_context: &SpatialSceneEntityContext<'_>,
     ) -> Result<(), QueryError> {
@@ -101,18 +102,14 @@ impl Points2DPart {
                 &instance_path_hashes_for_picking,
                 &colors,
                 &annotation_infos,
+                ent_context.world_from_obj,
             )?);
         }
 
         {
             let mut point_builder = ent_context.shared_render_builders.points();
             let point_batch = point_builder
-                .batch("2d points")
-                .depth_offset(ent_context.depth_offset)
-                .flags(
-                    re_renderer::renderer::PointCloudBatchFlags::FLAG_DRAW_AS_CIRCLES
-                        | re_renderer::renderer::PointCloudBatchFlags::FLAG_ENABLE_SHADING,
-                )
+                .batch("3d points")
                 .world_from_obj(ent_context.world_from_obj)
                 .outline_mask_ids(ent_context.highlight.overall)
                 .picking_object_id(re_renderer::PickingLayerObjectId(ent_path.hash64()));
@@ -121,14 +118,13 @@ impl Points2DPart {
                 re_tracing::profile_scope!("collect_points");
                 ent_view
                     .iter_primary()?
-                    .filter_map(|pt| pt.map(glam::Vec2::from))
+                    .filter_map(|pt| pt.map(glam::Vec3::from))
             };
 
             let picking_instance_ids = ent_view
                 .iter_instance_keys()
                 .map(picking_id_from_instance_key);
-
-            let mut point_range_builder = point_batch.add_points_2d(
+            let mut point_range_builder = point_batch.add_points(
                 ent_view.num_instances(),
                 point_positions,
                 radii,
@@ -153,14 +149,14 @@ impl Points2DPart {
                     }
                 }
             }
-        };
+        }
 
         load_keypoint_connections(ent_context, ent_path, keypoints);
 
         self.data.extend_bounding_box_with_points(
             ent_view
                 .iter_primary()?
-                .filter_map(|pt| pt.map(|pt| glam::vec3(pt.x, pt.y, 0.0))),
+                .filter_map(|pt| pt.map(|pt| pt.into())),
             ent_context.world_from_obj,
         );
 
@@ -168,10 +164,10 @@ impl Points2DPart {
     }
 }
 
-impl ViewPartSystem<SpatialSpaceView> for Points2DPart {
+impl ViewPartSystem<SpatialSpaceView> for Points3DPart {
     fn archetype(&self) -> ArchetypeDefinition {
         vec1::vec1![
-            Point2D::name(),
+            Point3D::name(),
             InstanceKey::name(),
             ColorRGBA::name(),
             Radius::name(),
@@ -189,9 +185,9 @@ impl ViewPartSystem<SpatialSpaceView> for Points2DPart {
         scene_context: &SpatialSceneContext,
         highlights: &SpaceViewHighlights,
     ) -> Vec<re_renderer::QueueableDrawData> {
-        re_tracing::profile_scope!("Points2DPart");
+        re_tracing::profile_scope!("Points3DPart");
 
-        process_entity_views::<re_components::Point2D, 7, _>(
+        process_entity_views::<re_components::Point3D, 7, _>(
             ctx,
             query,
             scene_context,
@@ -199,6 +195,9 @@ impl ViewPartSystem<SpatialSpaceView> for Points2DPart {
             scene_context.depth_offsets.points,
             self.archetype(),
             |_ctx, ent_path, entity_view, ent_context| {
+                scene_context
+                    .num_3d_primitives
+                    .fetch_add(1, std::sync::atomic::Ordering::Relaxed);
                 self.process_entity_view(query, &entity_view, ent_path, ent_context)
             },
         );
