+++ crates/re_space_view_spatial/src/parts/spatial_scene_part_data.rs
@@ -1,4 +1,4 @@
-use crate::scene::UiLabel;
+use crate::parts::UiLabel;
 
 /// Common data struct for all spatial scene elements.
 ///
